Update TeslaData Widget.js
Embedded a map key for consumer use. Store local copies of the map to avoid maxing out the "free" dev key (I should be able to support several hundred users for free, and it's unlikely we'll reach that point). Added links to Google Maps or Apple Maps when tapping on the map.